finish moving logic out of worker project. Removed unused dependencies
