+++ src/QueueReciver.Worker/Worker.cs
@@ -1,12 +1,7 @@
-using Microsoft.Azure.ServiceBus;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
 using QueueReceiver.Core.Interfaces;
-using QueueReceiver.Core.Models;
 using System;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -14,78 +9,28 @@ namespace QueueReceiverService
 {
     public class Worker : BackgroundService
     {
-        private readonly IQueueClient _queueClient;
-        private readonly IServiceScopeFactory _scopeFactory;
         private readonly ILogger<Worker> _logger;
         private readonly IEntryPointService entryPointService;
-        public Worker(ILogger<Worker> logger,
-            IServiceScopeFactory scopeFactory,
-            IQueueClient queueClient, 
-            IEntryPointService entryPointService)
+
+        public Worker(ILogger<Worker> logger, IEntryPointService entryPointService)
         {
             _logger = logger;
-            _scopeFactory = scopeFactory;
-            _queueClient = queueClient;
-            RegisterOnMessageHandlerAndReceiveMessages();
             this.entryPointService = entryPointService;
-        }
-
-        private void RegisterOnMessageHandlerAndReceiveMessages()
-        {
-            var messageHandlerOptions = new MessageHandlerOptions(ExceptionReceivedHandler)
-            {
-                MaxConcurrentCalls = 1,
-                AutoComplete = false
-            };
-            _queueClient.RegisterMessageHandler(ProcessMessagesAsync, messageHandlerOptions);
-        }
-
-        private async Task ProcessMessagesAsync(Message message, CancellationToken token)
-        {
-            var accessInfo = JsonConvert.DeserializeObject<AccessInfo>(Encoding.UTF8.GetString(message.Body));
-            _logger.LogInformation($"Processing message : { accessInfo }");
-
-            /**
-             * Injecting here because Worker is singleton.
-             * It not possible initiate scoped dependiences from a constructor of a singleton  
-            **/
-            using var scope = _scopeFactory.CreateScope();
-                var accessService =
-                    scope.ServiceProvider
-                        .GetRequiredService<IAccessService>();
-
-            await accessService.HandleRequest(accessInfo);
-            //TODO figure out exception handleing
-            //TODO consider moving all logic out of this project
-
-            await _queueClient.CompleteAsync(message.SystemProperties.LockToken);
-            _logger.LogInformation($"Message completed successfully");
-        }
-
-        private Task ExceptionReceivedHandler(ExceptionReceivedEventArgs exceptionReceivedEventArgs)
-        {
-            _logger.LogError(exceptionReceivedEventArgs.Exception, "Message handler encountered an exception");
-            var context = exceptionReceivedEventArgs.ExceptionReceivedContext;
-
-            _logger.LogDebug($"- Endpoint: {context.Endpoint}");
-            _logger.LogDebug($"- Entity Path: {context.EntityPath}");
-            _logger.LogDebug($"- Executing Action: {context.Action}");
-            return Task.CompletedTask;
-        }
-
-        public async override Task StartAsync(CancellationToken cancellationToken) 
-        {
-            await entryPointService.ExcecuteAsync();
         }
 
         protected override async Task ExecuteAsync(CancellationToken stoppingToken)
         {
+            _logger.LogInformation("Worker service at: {time}", DateTimeOffset.Now);
+            await entryPointService.InitializeQueue();
+
             while (!stoppingToken.IsCancellationRequested)
             {
                 _logger.LogInformation("Worker running at: {time}", DateTimeOffset.Now);
                 await Task.Delay(100000000, stoppingToken);
             }
-            await _queueClient.CloseAsync();
+
+            await entryPointService.DisposeQueue();
+            _logger.LogInformation("Worker service stopping at: {time}", DateTimeOffset.Now);
         }
     }
 }
